+++ morpheus-graphql-code-gen/morpheus-graphql-code-gen.cabal
@@ -35,11 +35,9 @@ library
       Data.Morpheus.CodeGen.Server.Printing.GQLType
       Data.Morpheus.CodeGen.Server.Printing.Render
       Data.Morpheus.CodeGen.Server.Printing.Type
-      Data.Morpheus.CodeGen.Server.TH.Compile
       Data.Morpheus.CodeGen.Server.TH.Declare
       Data.Morpheus.CodeGen.Server.TH.GQLDirective
       Data.Morpheus.CodeGen.Server.TH.GQLType
-      Data.Morpheus.CodeGen.Server.TH.Type
       Data.Morpheus.CodeGen.Server.TH.Utils
       Paths_morpheus_graphql_code_gen
   hs-source-dirs:
