--- morpheus-graphql-code-gen/src/Data/Morpheus/CodeGen/Server/TH/Compile.hs
@@ -1,44 +1,0 @@
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE NoImplicitPrelude #-}
-
-module Data.Morpheus.CodeGen.Server.TH.Compile
-  ( compileDocument,
-    gqlDocument,
-  )
-where
-
---
---  Morpheus
-
-import Data.ByteString.Lazy.Char8 (ByteString)
-import qualified Data.ByteString.Lazy.Char8 as LB
-import Data.Morpheus.CodeGen.Server.Internal.AST
-  ( CodeGenConfig (..),
-  )
-import Data.Morpheus.CodeGen.Server.Interpreting.Transform
-  ( parseServerTypeDefinitions,
-  )
-import Data.Morpheus.CodeGen.Server.TH.Declare
-  ( runDeclare,
-  )
-import Language.Haskell.TH
-import Language.Haskell.TH.Quote (QuasiQuoter (..))
-import Relude hiding (ByteString)
-
-gqlDocument :: QuasiQuoter
-gqlDocument = mkQuasiQuoter CodeGenConfig {namespace = False}
-
-mkQuasiQuoter :: CodeGenConfig -> QuasiQuoter
-mkQuasiQuoter ctx =
-  QuasiQuoter
-    { quoteExp = notHandled "Expressions",
-      quotePat = notHandled "Patterns",
-      quoteType = notHandled "Types",
-      quoteDec = compileDocument ctx . LB.pack
-    }
-  where
-    notHandled things =
-      error $ things <> " are not supported by the GraphQL QuasiQuoter"
-
-compileDocument :: CodeGenConfig -> ByteString -> Q [Dec]
-compileDocument ctx = parseServerTypeDefinitions ctx >=> runDeclare ctx
